# FP1hacktiv8/Project-1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Hacktiv8 Final Project 1 - Group 1",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/todos": {
            "get": {
                "description": "Get All Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get All Data Todos",
                "responses": {}
            },
            "post": {
                "description": "Post New Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Post New Data Todos",
                "parameters": [
                    {
                        "description": "Todos",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessCreate"
                        }
                    }
                }
            }
        },
        "/todos/delete/{id}": {
            "delete": {
                "description": "Delete Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Data Todos",
                "parameters": [
                    {
                        "description": "Delete",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessDelete"
                        }
                    }
                }
            }
        },
        "/todos/update/{id}": {
            "put": {
                "description": "Edit Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Edit Data Todos",
                "parameters": [
                    {
                        "description": "Update",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessUpdate"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Get Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Data Todos",
                "parameters": [
                    {
                        "description": "Todos",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Todo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tolist": {
                    "type": "string"
                }
            }
        },
        "response.SuccessCreate": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Todo"
                },
                "messages": {
                    "type": "string",
                    "example": "success add new todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessDelete": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success delete todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessUpdate": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success update todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}